+++ src/Field.php
@@ -19,6 +19,7 @@ use benf\neo\elements\db\BlockQuery;
 use benf\neo\models\BlockStructure;
 use benf\neo\models\BlockType;
 use benf\neo\models\BlockTypeGroup;
+use benf\neo\validators\FieldValidator;
 
 /**
  * Class Field
@@ -62,6 +63,21 @@ class Field extends BaseField implements EagerLoadingFieldInterface
 	public $localizeBlocks = false;
 
 	/**
+	 * @var int|null The minimum number of blocks this field can have.
+	 */
+	public $minBlocks;
+
+	/**
+	 * @var int|null The maximum number of blocks this field can have.
+	 */
+	public $maxBlocks;
+
+	/**
+	 * @var int|null The maximum number of top-level blocks this field can have.
+	 */
+	public $maxTopBlocks;
+
+	/**
 	 * @var array|null The block types associated with this field.
 	 */
 	private $_blockTypes;
@@ -77,20 +93,10 @@ class Field extends BaseField implements EagerLoadingFieldInterface
 	public function rules(): array
 	{
 		$rules = parent::rules();
-		$rules[] = [['minBlocks', 'maxBlocks'], 'integer', 'min' => 0];
+		$rules[] = [['minBlocks', 'maxBlocks', 'maxTopBlocks'], 'integer', 'min' => 0];
 
 		return $rules;
 	}
-
-	/**
-	 * @var int|null The minimum number of blocks this field can have.
-	 */
-	public $minBlocks;
-
-	/**
-	 * @var int|null The maximum number of blocks this field can have.
-	 */
-	public $maxBlocks;
 
 	/**
 	 * Returns this field's block types.
@@ -439,6 +445,11 @@ class Field extends BaseField implements EagerLoadingFieldInterface
 				'tooFew' => Craft::t('neo', '{attribute} should contain at least {min, number} {min, plural, one{block} other{blocks}}.'),
 				'tooMany' => Craft::t('neo', '{attribute} should contain at most {max, number} {max, plural, one{block} other{blocks}}.'),
 				'skipOnEmpty' => false,
+				'on' => Element::SCENARIO_LIVE,
+			],
+			[
+				FieldValidator::class,
+				'maxTopBlocks' => $this->maxTopBlocks ?: null,
 				'on' => Element::SCENARIO_LIVE,
 			],
 		];