+++ src/Plugin.php
@@ -5,12 +5,16 @@ use yii\base\Event;
 
 use Craft;
 use craft\base\Plugin as BasePlugin;
+use craft\db\Query;
+use craft\events\RebuildConfigEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\services\Fields;
+use craft\services\ProjectConfig;
 use craft\web\twig\variables\CraftVariable;
 
 use benf\neo\controllers\Conversion as ConversionController;
 use benf\neo\controllers\Input as InputController;
+use benf\neo\integrations\fieldlabels\FieldLabels;
 use benf\neo\models\Settings;
 use benf\neo\services\Blocks as BlocksService;
 use benf\neo\services\BlockTypes as BlockTypesService;
@@ -81,13 +85,15 @@ class Plugin extends BasePlugin
 			}
 		);
 
-		Craft::$app->getProjectConfig()
-			->onAdd('neoBlockTypes.{uid}', [$this->blockTypes, 'handleChangedBlockType'])
-			->onUpdate('neoBlockTypes.{uid}', [$this->blockTypes, 'handleChangedBlockType'])
-			->onRemove('neoBlockTypes.{uid}', [$this->blockTypes, 'handleDeletedBlockType'])
-			->onAdd('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleChangedBlockTypeGroup'])
-			->onUpdate('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleChangedBlockTypeGroup'])
-			->onRemove('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleDeletedBlockTypeGroup']);
+		// Setup project config functionality
+		$this->_setupProjectConfig();
+
+		$pluginsService = Craft::$app->getPlugins();
+
+		if ($pluginsService->isPluginInstalled('fieldlabels'))
+		{
+			(new FieldLabels)->init();
+		}
 
 		if (class_exists('\NerdsAndCompany\Schematic\Schematic')) {
 			Event::on(
@@ -110,5 +116,97 @@ class Plugin extends BasePlugin
 	protected function createSettingsModel(): Settings
 	{
 		return new Settings();
+	}
+
+	private function _setupProjectConfig()
+	{
+		// Listen for Neo updates in the project config to apply them to the database
+		Craft::$app->getProjectConfig()
+			->onAdd('neoBlockTypes.{uid}', [$this->blockTypes, 'handleChangedBlockType'])
+			->onUpdate('neoBlockTypes.{uid}', [$this->blockTypes, 'handleChangedBlockType'])
+			->onRemove('neoBlockTypes.{uid}', [$this->blockTypes, 'handleDeletedBlockType'])
+			->onAdd('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleChangedBlockTypeGroup'])
+			->onUpdate('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleChangedBlockTypeGroup'])
+			->onRemove('neoBlockTypeGroups.{uid}', [$this->blockTypes, 'handleDeletedBlockTypeGroup']);
+
+		// Listen for a project config rebuild, and provide the Neo data from the database
+		Event::on(ProjectConfig::class, ProjectConfig::EVENT_REBUILD, function(RebuildConfigEvent $event)
+		{
+			$fieldsService = Craft::$app->getFields();
+			$blockTypeData = [];
+			$blockTypeGroupData = [];
+
+			$blockTypeQuery = (new Query)
+				->select([
+					// We require querying for the layout ID, rather than performing an inner join and getting the
+					// layout UID that way, because Neo allows block types not to have field layouts
+					'types.fieldLayoutId',
+					'types.name',
+					'types.handle',
+					'types.maxBlocks',
+					'types.maxChildBlocks',
+					'types.childBlocks',
+					'types.topLevel',
+					'types.sortOrder',
+					'types.uid',
+					'fields.uid AS field',
+				])
+				->from(['{{%neoblocktypes}} types'])
+				->innerJoin('{{%fields}} fields', '[[types.fieldId]] = [[fields.id]]');
+
+			foreach ($blockTypeQuery->all() as $blockType)
+			{
+				$childBlocks = $blockType['childBlocks'];
+
+				if (!empty($childBlocks))
+				{
+					$childBlocks = json_decode($childBlocks);
+				}
+
+				$blockTypeData[$blockType['uid']] = [
+					'field' => $blockType['field'],
+					'name' => $blockType['name'],
+					'handle' => $blockType['handle'],
+					'sortOrder' => (int)$blockType['sortOrder'],
+					'maxBlocks' => (int)$blockType['maxBlocks'],
+					'maxChildBlocks' => (int)$blockType['maxChildBlocks'],
+					'childBlocks' => $childBlocks,
+					'topLevel' => (bool)$blockType['topLevel'],
+				];
+
+				if ($blockType['fieldLayoutId'] !== null)
+				{
+					$fieldLayout = $fieldsService->getLayoutById($blockType['fieldLayoutId']);
+					$fieldLayoutConfig = $fieldLayout->getConfig();
+					$blockType['fieldLayouts'] = [
+						$fieldLayout->uid => $fieldLayoutConfig,
+					];
+				}
+
+				unset($blockType['fieldLayoutId']);
+			}
+
+			$blockTypeGroupQuery = (new Query())
+				->select([
+					'groups.name',
+					'groups.sortOrder',
+					'groups.uid',
+					'fields.uid AS field',
+				])
+				->from(['{{%neoblocktypegroups}} groups'])
+				->innerJoin('{{%fields}} fields', '[[groups.fieldId]] = [[fields.id]]');
+
+			foreach ($blockTypeGroupQuery->all() as $blockTypeGroup)
+			{
+				$blockTypeGroupData[$blockTypeGroup['uid']] = [
+					'field' => $blockTypeGroup['field'],
+					'name' => $blockTypeGroup['name'],
+					'sortOrder' => $blockTypeGroup['sortOrder'],
+				];
+			}
+
+			$event->config['neoBlockTypes'] = $blockTypeData;
+			$event->config['neoBlockTypeGroups'] = $blockTypeGroupData;
+		});
 	}
 }